+++ src/Symfony/Component/Console/Helper/ProgressBar.php
@@ -49,6 +49,7 @@ final class ProgressBar
     private float $maxSecondsBetweenRedraws = 1;
     private OutputInterface $output;
     private int $step = 0;
+    private int $resumedStep = 0;
     private ?int $max = null;
     private int $startTime;
     private int $stepWidth;
@@ -199,11 +200,11 @@ final class ProgressBar
 
     public function getEstimated(): float
     {
-        if (!$this->step) {
+        if (0 === $this->step || $this->step === $this->resumedStep) {
             return 0;
         }
 
-        return round((time() - $this->startTime) / $this->step * $this->max);
+        return round((time() - $this->startTime) / ($this->step - $this->resumedStep) * $this->max);
     }
 
     public function getRemaining(): float
@@ -212,7 +213,7 @@ final class ProgressBar
             return 0;
         }
 
-        return round((time() - $this->startTime) / $this->step * ($this->max - $this->step));
+        return round((time() - $this->startTime) / ($this->step - $this->resumedStep) * ($this->max - $this->step));
     }
 
     public function setBarWidth(int $size)
@@ -302,13 +303,17 @@ final class ProgressBar
     /**
      * Starts the progress output.
      *
-     * @param int|null $max Number of steps to complete the bar (0 if indeterminate), null to leave unchanged
+     * @param int|null $max      Number of steps to complete the bar (0 if indeterminate), null to leave unchanged
+     * @param int      $resumeAt when restarting a previously started progress, set on what step we are restarting so that time estimations are calculated correctly, and the
+     *                           progress is automatically set to the appropriate step
      */
-    public function start(int $max = null)
+    public function start(int $max = null, int $resumeAt = 0): void
     {
         $this->startTime = time();
-        $this->step = 0;
-        $this->percent = 0.0;
+        $this->step = $resumeAt;
+        $this->resumedStep = $resumeAt;
+
+        $resumeAt > 0 ? $this->setProgress($resumeAt) : $this->percent = 0.0;
 
         if (null !== $max) {
             $this->setMaxSteps($max);